\name{all_equal}
\alias{all_equal}
\title{Checks if elements are all equal}
\description{
The function checks if all elements in a vector are equal.
}
\usage{
all_equal(x, tol = .Machine$double.eps ^ 0.5)
}
\arguments{
\item{x}{The vector to be checked.}
\item{tol}{The tolerance.}
}
\value{
\item{x}{A logical value.}
}

