% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_DD}
\alias{build_DD}
\title{Build the \eqn{D} matrix}
\usage{
build_DD(d, n_lags)
}
\arguments{
\item{d}{The matrix of size \code{(n_T + n_lags) * n_determ} of deterministic terms.}

\item{n_lags}{The number of lags.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
\item{DD}{A matrix of size \code{n_T * ((n_lags + 1)*n_determ)} where row \code{t} is \eqn{(d_t', -d_{t-1}', \dots, -d_{t-k}')}.}
}
\description{
\code{build_DD} builds the \eqn{D} matrix.
}
