% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/builders.R
\name{build_U_cpp}
\alias{build_U_cpp}
\alias{build_U}
\title{Template titel}
\usage{
build_U_cpp(Pi, n_determ, n_vars, n_lags)

build_U(Pi, n_determ)
}
\arguments{
\item{Pi}{Matrix of size \code{n_vars * (n_vars*n_lags)} containing the dynamic coefficients.}

\item{n_determ}{The number of deterministic terms.}

\item{n_vars}{The number of variables.}

\item{n_lags}{The number of lags.}
}
\value{
\item{U}{The \code{U} matrix, of size \code{((n_lags+1)n_vars*n_determ) * n_vars*n_determ}.}
}
\description{
Builds the parameter matrix of dynamic coefficients for the companion form representation.
}
\section{Functions}{
\itemize{
\item \code{build_U_cpp}: Build the U matrix (C++ implementation)

\item \code{build_U}: Build the U matrix (R implementation)
}}

