% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_params}
\alias{check_required_params}
\title{Check required parameters}
\usage{
check_required_params(x, ...)
}
\arguments{
\item{x}{prior object}

\item{...}{names of parameters whose existence is to be checked}
}
\description{
Check whether prior contains the required parameters
}
