% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_error_variances}
\alias{compute_error_variances}
\title{Compute error variances}
\usage{
compute_error_variances(Y)
}
\arguments{
\item{Y}{data matrix}
}
\value{
vector with error variances
}
\description{
Function to run independent AR(4) regressions to obtain preliminary estimates of error variances
}
