% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pi_sigma.R
\name{create_prior_Pi_Omega}
\alias{create_prior_Pi_Omega}
\title{Create the priors for Pi and Sigma}
\usage{
create_prior_Pi_Omega(lambda1, lambda2, lambda3, prior_Pi_AR1, Y, n_lags)
}
\arguments{
\item{lambda1}{overall tightness}

\item{lambda2}{cross-equation tightness}

\item{lambda3}{lag decay}

\item{prior_Pi_AR1}{prior means for AR(1) coefficients}

\item{Y}{data}

\item{n_lags}{number of lags}
}
\value{
\item{prior_Pi_Omega}{The prior covariance matrix for Pi.}
}
\description{
Creates the prior mean and covariance for Pi given the hyperparameters, and the prior parameters for Sigma.
}
\keyword{internal}
