% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_list}
\alias{data_list}
\title{Real-time data set.}
\format{A list with three components of length 192.
\describe{
  \item{fcst_date}{The date at which this data was available}
  \item{mf}{The monthly data set (with missingness)}
  \item{qf}{The quarterly data set}
}}
\usage{
data_list
}
\description{
A dataset containing real-time data for mixed and quarterly frequencies.
}
\keyword{datasets}
