% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dmatt}
\alias{dmatt}
\title{Matrix t distribution}
\usage{
dmatt(X, M, P, Q, v)
}
\arguments{
\item{X}{\code{p * q} matrix at which the density is to be evaluated}

\item{M}{\code{p * q} matrix of means}

\item{P}{\code{p * p} scale matrix}

\item{Q}{\code{q * q} scale matrix}

\item{v}{degrees of freedom}
}
\value{
For \code{dmultn}: the evaluated density.\\n
For \code{rmultn}: \eqn{p} random numbers.
}
\description{
Density function for the truncated multivariate normal distribution
}
\references{
Karlsson, S. (2013) Forecasting with Bayesian Vector Autoregression.
In Elliott, G. and Timmermann, A., editors, \emph{Handbook of Economic Forecasting},
volume 2, chapter 15, pp. 791-897. Elsevier B.V.
}
\keyword{internal}
