% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dmultn}
\alias{dmultn}
\title{Multivariate normal density function}
\usage{
dmultn(x, m, Sigma)
}
\arguments{
\item{x}{A vector of size \code{p}.}

\item{m}{The mean vector of size \code{p}.}

\item{Sigma}{The covariance matrix.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
The evaluated density.
}
\description{
Density function for the multivariate normal distribution
}
