% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dnorminvwish}
\alias{dnorminvwish}
\title{Normal inverse Wishart density function}
\usage{
dnorminvwish(X, Sigma, M, P, S, v)
}
\arguments{
\item{X}{Matrix of size \code{p * q}.}

\item{Sigma}{The covariance matrix.}

\item{M}{The mean matrix of size \code{p * q}.}

\item{P}{\code{p * p} covariance matrix.}

\item{S}{\code{q * q} scale matrix.}

\item{v}{The degrees of freedom.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
The evaluated density.
}
\description{
Density function for the (matrix) normal inverse Wishart distribution
}
