% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_info.R
\name{get_params_info}
\alias{get_params_info}
\title{Get info about parameters}
\usage{
get_params_info(
  minn = FALSE,
  ss = FALSE,
  ssng = FALSE,
  dl = FALSE,
  iw = FALSE,
  diffuse = FALSE,
  csv = FALSE,
  fsv = FALSE
)
}
\arguments{
\item{minn}{logical for \code{minn} prior}

\item{ss}{logical for \code{ss} prior}

\item{ssng}{logical for \code{ssng} prior}

\item{dl}{logical for \code{dl} prior}

\item{iw}{logical for \code{iw} prior}

\item{diffuse}{logical for \code{diffuse} prior}

\item{csv}{logical for \code{csv} prior}

\item{fsv}{logical for \code{fsv} prior}
}
\value{
A list with components:
\item{prior_params}{Components in the prior object that need to be extracted}
\item{required_params}{Components that must be non-null}
\item{retrieved_params}{Components that are retrieved in initialization}
\item{params}{Parameters that are sampled}
}
\description{
Function that provides character vectors indicating different types of
parameters needed for a certain model
}
