% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_minn.R
\name{gibbs_sampler_minn}
\alias{gibbs_sampler_minn}
\title{Gibbs sampler for Mixed-Frequency BVAR}
\usage{
gibbs_sampler_minn(Y, freq, prior_Pi_AR1, lambda1, lambda2, lambda3, n_lags,
  n_fcst = NULL, n_reps, init_Pi = NULL, init_Sigma = NULL,
  init_Z = NULL, smooth_state = FALSE, check_roots = TRUE,
  verbose = TRUE)
}
\arguments{
\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}

\item{freq}{Character vector with elements 'm' (monthly) or 'q' (quarterly) for sampling frequency.}

\item{prior_Pi_AR1}{The prior means for the AR(1) coefficients.}

\item{lambda1}{The overall tightness.}

\item{lambda2}{The lag decay.}

\item{lambda3}{The tightness of the intercept prior variance.}

\item{n_lags}{The number of lags.}

\item{n_fcst}{The number of periods to forecast.}

\item{n_reps}{The number of replications.}

\item{init_Pi}{Matrix to initialize the dynamic coefficients.}

\item{init_Sigma}{Matrix to initialize the error covariance.}

\item{init_Z}{Matrix to initialize the latent state.}

\item{smooth_state}{Logical, if \code{TRUE} then the smoothed estimates of the latent states are also returned.}

\item{check_roots}{Logical, if roots of the companion matrix are to be checked to ensure stationarity.}

\item{verbose}{Logical, if progress should be printed to the console.}
}
\value{
An object of class mfbvar.
}
\description{
\code{gibbs_sampler} runs a Gibbs sampler to approximate the posterior of the VAR model parameters.
}
\details{
The prior covariance of \eqn{\Pi} given \eqn{\Sigma} is \eqn{\Sigma \otimes \Omega_\Pi}.

The function \code{gibbs_sampler_qf} is a temporary version of \code{gibbs_sampler} which is customized for quarterly data (for improved speed). Its purpose is to accomomdate conditional forecasting. The function \code{gibbs_sampler2} uses an R implementation of the simulation smoother, whereas \code{gibbs_sampler} uses a C++ implementation.
}
\keyword{internal}
