% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kf_smoother}
\alias{kf_smoother}
\title{Kalman smoother with correlated errors}
\usage{
kf_smoother(y, Z, c, G, T, d, H, a1, P1)
}
\arguments{
\item{y}{(\code{T x n}) data matrix (\code{NA} is allowed)}

\item{Z}{(\code{n x s}) matrix relating state to measurement}

\item{c}{(\code{T x n}) matrix of deterministic variable(s) in measurement equation}

\item{G}{(\code{n x n}) matrix of (matrix) square root of error covariance matrix in measurement equation}

\item{T}{(\code{s x s}) matrix relating state at \eqn{t-1} to state at \eqn{t}}

\item{d}{(\code{T x s}) matrix of deterministic variable(s) in state equation}

\item{H}{(\code{s x n}) matrix of (matrix) square root of error covariance matrix in state equation}

\item{a1}{(\code{s x 1}) matrix of initial value for the state}

\item{P1}{(\code{s x s}) matrix of initial covariance for the state}
}
\value{
A list with elements:
\describe{
\item{\code{a}}{one-step predictions (where \code{a[t]} corresponds to \eqn{a_t})}
\item{\code{a_tt}}{filtered estimates (where \code{a_tt[t]} corresponds to \eqn{a_{t|t}})}
\item{\code{a_tT}}{smoothed estimates (where \code{a_tT[t]} corresponds to \eqn{a_{t|T}})}
}
}
\description{
This function performs Kalman smoothing for a state-space model with possibly correlated errors in the measurement and state equations.
}
\seealso{
\code{\link[smoothr]{kf_filter}}.
}
