% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_variables}
\alias{list_to_variables}
\title{List to variables}
\usage{
list_to_variables(x, envir, ...)
}
\arguments{
\item{x}{list}

\item{envir}{environment}

\item{...}{variables existing in the list that are to be assigned to \code{envir}}
}
\description{
Transform a list to variables, assigning variables to the given environment
}
