\name{max_eig_cpp}
\alias{max_eig_cpp}
\alias{max_eig_cpp}
\title{Find maximum eigenvalue}
\description{
The function computes the maximum eigenvalue.
}
\usage{
max_eig_cpp(M)
}
\arguments{
\item{M}{The symmetric matrix to compute the maximum eigenvalue of.}
}
\value{
\item{x}{The maximum eigenvalue.}
}

