% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_eig_cpp}
\alias{max_eig_cpp}
\alias{max_eig}
\title{Find maximum eigenvalue}
\usage{
max_eig_cpp(A)
}
\arguments{
\item{A}{Symmetrix matrix whose maximum eigenvalue is to be computed.}
}
\value{
The maximum eigenvalue.
}
\description{
The function computes the maximum eigenvalue.
}
