% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_minn_diffuse}
\alias{mcmc_minn_diffuse}
\title{MCMC Sampler for Minnesota with a Diffuse Error Covariance Prior}
\usage{
mcmc_minn_diffuse(y_in_p, Pi, Sigma, Z, Z_fcst, aux, global, local, slice,
  Lambda_comp, prior_Pi_Omega, prior_Pi_mean_vec, Z_1, n_reps, n_burnin,
  n_q, T_b, n_lags, n_vars, n_T, n_fcst, n_thin, verbose, a, gig)
}
\arguments{
\item{y_in_p}{Matrix of data}

\item{Pi}{Array for storing regression parameters}

\item{Sigma}{Array for storing the error covariance matrix}

\item{Z}{Array for storing the data}

\item{Z_fcst}{Array for storing the forecasts}

\item{aux}{Matrix for storing the auxiliary hyperparameters in the Dirichlet-Laplace prior}

\item{global}{Vector for storing the global hyperparameters in the Dirichlet-Laplace prior}

\item{local}{Matrix for storing the local hyperparameters in the Dirichlet-Laplace prior}

\item{slice}{Vector that will hold the temporary slice variables if slice sampling is used for the Dirichet-Laplace prior}

\item{Lambda_comp}{Aggregation matrix}

\item{prior_Pi_Omega}{Matrix with prior variances of the regression parameters}

\item{prior_Pi_mean_vec}{Vector of prior means of the regression parameters}
}
\description{
MCMC Sampler for Minnesota with a Diffuse Error Covariance Prior
}
