% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{mcmc_sampler.mfbvar_ss}
\alias{mcmc_sampler.mfbvar_ss}
\alias{mcmc_sampler.mfbvar_minn}
\title{MCMC sampler for mixed-frequency BVAR}
\usage{
\method{mcmc_sampler}{mfbvar_ss}(x, ...)

\method{mcmc_sampler}{mfbvar_minn}(x, ...)
}
\arguments{
\item{x}{a prior object (inhereting from \code{mfbvar_prior})}

\item{...}{additional arguments (\code{n_reps} and \code{init})}
}
\value{
An object of class \code{mfbvar} and \code{mfbvar_ss} or \code{mfbvar_minn}.
}
\description{
MCMC sampler to approximate the posterior distribution of the VAR model
parameters.
}
\details{
\code{mcmc_sampler.mfbvar_ss} is used for a steady-state prior and
\code{mcmc_sampler.mfbvar_minn} for a Minnesota prior
}
\keyword{internal}
