% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R, R/mcmc_sampler_sv.R
\name{mcmc_sampler}
\alias{mcmc_sampler}
\alias{mcmc_sampler.mfbvar_ss_iw}
\alias{mcmc_sampler.mfbvar_minn_iw}
\alias{mcmc_sampler.mfbvar_minn_fsv}
\alias{mcmc_sampler.mfbvar_ss_fsv}
\title{MCMC sampler}
\usage{
mcmc_sampler(x, ...)

\method{mcmc_sampler}{mfbvar_ss_iw}(x, ...)

\method{mcmc_sampler}{mfbvar_minn_iw}(x, ...)

\method{mcmc_sampler}{mfbvar_minn_fsv}(x, ...)

\method{mcmc_sampler}{mfbvar_ss_fsv}(x, ...)
}
\arguments{
\item{x}{argument to dispatch on (of class \code{prior_obj})}

\item{...}{additional named arguments passed on to the methods}
}
\description{
\code{mcmc_sampler} is a generic function for deciding which specific MCMC
algorithm to dispatch to. It is called internally.
}
