% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_minn_iw}
\alias{mdd.mfbvar_minn_iw}
\title{Marginal data density method for class \code{mfbvar_minn}}
\usage{
\method{mdd}{mfbvar_minn_iw}(x, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_minn}}

\item{...}{additional arguments (currently only \code{p_trunc} for the degree of truncation is available)}
}
\value{
The logarithm of the marginal data density.
}
\description{
Estimate the marginal data density for the model with a Minnesota prior.
}
\details{
The method used for estimating the marginal data density is the proposal made by
Schorfheide and Song (2015).
}
\references{
Schorfheide, F., & Song, D. (2015) Real-Time Forecasting With a Mixed-Frequency VAR. \emph{Journal of Business & Economic Statistics}, 33(3), 366--380. \doi{10.1080/07350015.2014.954707}
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_ss_iw}}
}
