% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_ss}
\alias{mdd.mfbvar_ss}
\title{Marginal data density method for class \code{mfbvar_ss}}
\usage{
\method{mdd}{mfbvar_ss}(x, method = 1, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_ss}}

\item{method}{option for which method to choose for computing the mdd}

\item{...}{additional arguments (currently only \code{p_trunc} for the degree of truncation for method 2 is available)}
}
\description{
Estimate the marginal data density for the model with a steady-state prior.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_minn}}
}
