% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.minn}
\alias{mdd.minn}
\alias{mdd_minn}
\title{Marginal data density}
\usage{
\method{mdd}{minn}(mfbvar_obj, quarterly_cols, type = "full", ...)

mdd_minn(mfbvar_obj, quarterly_cols, type = "full", ...)
}
\arguments{
\item{mfbvar_obj}{An object of class \code{mfbvar} containing the results.}
}
\value{
The log marginal data density estimate (bar a constant)
}
\description{
Estimate the marginal data density for the model with a Minnesota prior.

This function provides the possibility to estimate the log marginal density (up to a constant) using the Schorfheide-Song MFBVAR.
}
\details{
The generic \code{mdd} calls \code{mdd_minn} if \code{mfbvar_obj} is of class \code{mfbvar_minn}.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.ss}}
}
\keyword{internal}
