% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.ss}
\alias{mdd.ss}
\alias{mdd1}
\alias{mdd2}
\title{Marginal data density}
\usage{
\method{mdd}{ss}(mfbvar_obj, method = c(1, 2), ...)

mdd1(mfbvar_obj)

mdd2(mfbvar_obj, p_trunc)
}
\arguments{
\item{mfbvar_obj}{An object of class \code{mfbvar} containing the results.}

\item{method}{option for which method to choose for computing the mdd}

\item{...}{additional arguments (currently only \code{p_trunc} available)}

\item{p_trunc}{\code{1-p_trunc} is the degree of truncation (i.e. \code{p_trunc=1} is no truncation).}
}
\value{
\code{mdd1} returns a list with components (all are currently in logarithms):
\item{lklhd}{The likelihood.}
\item{eval_prior_Pi_Sigma}{The evaluated prior.}
\item{eval_prior_psi}{The evaluated prior of psi.}
\item{eval_RB_Pi_Sigma}{The Rao-Blackwellized estimate of the conditional posterior of Pi and Sigma.}
\item{eval_marg_psi}{The evaluated marginal posterior of psi.}
\item{log_mdd}{The mdd estimate (in log).}

\code{mdd} returns a list with components being \code{n_reps}-long vectors and a scalar (the final estimate).
\item{eval_posterior_Pi_Sigma}{Posterior of Pi and Sigma.}
\item{data_likelihood}{The likelihood.}
\item{eval_prior_Pi_Sigma}{Prior of Pi and Sigma.}
\item{eval_prior_psi}{Prior of psi.}
\item{psi_truncated}{The truncated psi pdf.}
\item{log_mdd}{The mdd estimate (in log).}
}
\description{
Estimate the marginal data density for the model with a steady-state prior.
}
\details{
The generic \code{mdd} calls either \code{mdd1} or \code{mdd2} if \code{mfbvar_obj} is of class \code{mfbvar_ss}.

\code{mdd1} uses method 1, \code{mdd2} uses method 2.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.minn}}
}
\keyword{internal}
