% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mdd1}
\alias{mdd1}
\title{Marginal data density (method 1)}
\usage{
mdd1(mfbvar_obj)
}
\arguments{
\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
A list with components:
\item{lklhd}{The likelihood.}
\item{Pi_Sigma_prior}{The evaluated prior.}
\item{psi_prior}{The evaluated prior of psi.}
\item{Pi_Sigma_RB}{The Rao-Blackwellized estimate of the conditional posterior of Pi and Sigma.}
\item{psi_MargPost}{The evaluated marginal posterior of psi.}
\item{mdd_estimate}{The final estimate.}
}
\description{
Estimate the marginal data density using method 1.
}
