% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd_grid}
\alias{mdd_grid}
\title{Marginal data density grid search}
\usage{
mdd_grid(mfbvar_obj = NULL, lambda1_grid, lambda2_grid, method, n_cores = 1,
  p_trunc = NULL, ...)
}
\arguments{
\item{mfbvar_obj}{An object of class \code{mfbvar} containing the results.}

\item{lambda1_grid}{The grid of values to use for lambda1.}

\item{lambda2_grid}{The grid of values to use for lambda2.}

\item{method}{The method to use for estimation of the log marginal data density. One of \code{1} and \code{2}.}

\item{n_cores}{The number of cores to use (if set to 1, computation is done serially).}

\item{p_trunc}{\code{1-p_trunc} is the degree of truncation (i.e. \code{p_trunc=1} is no truncation).}

\item{...}{Additional parameters. This includes \code{p_trunc} if \code{method == 2} and if \code{mfbvar_obj} is omitted, it must include all the necessary variables to the \code{mfbvar()} call.}
}
\value{
mdd_res A matrix with the results.
}
\description{
Use a grid search to estimate the marginal data density for multiple values, possibly using parallel computation.
}
\details{
Only one of \code{mfbvar_obj} and the entire suite of \code{Y}, \code{d}, \code{Lambda}, \code{prior_Pi_AR1}, \code{prior_nu}, \code{prior_psi_mean}, \code{prior_psi_Omega}, \code{n_lags}, \code{n_burnin} and \code{n_reps} must be provided. If \code{mfbvar_obj} is not \code{NULL}, its components are extracted.\cr \cr In order to use \code{n_cores > 1}, the \code{parallel} package must be installed.
}
