% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_schorf.R
\name{mdd_schorf}
\alias{mdd_schorf}
\title{Estimate MDD in Schorfheide-Song MFBVAR}
\usage{
mdd_schorf(monthly_cols, postsim, Z, n_T)
}
\arguments{
\item{n_T}{The number of time points.}
}
\value{
The log marginal data density estimate (bar a constant)
}
\description{
This function provides the possibility to estimate the log marginal density (up to a constant) using the Schorfheide-Song MFBVAR.
}
