% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf_sweden}
\alias{mf_sweden}
\title{Real-time data set.}
\format{A mixed-frequency data set of five Swedish macroeconomic variables.
\describe{
  \item{unemp}{harmonized unemployment rate (source: OECD)}
  \item{infl}{inflation rate (source: OECD)}
  \item{ip}{industrial production (source: OECD)}
  \item{eti}{economic tendency indicator (source: National Institute of Economic Research)}
  \item{gdp}{GDP growth (source: Statistics Sweden)}
}}
\usage{
mf_sweden
}
\description{
A dataset containing real-time data for mixed and quarterly frequencies.
}
\keyword{datasets}
