% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf_usa}
\alias{mf_usa}
\title{US Macroeconomic Data Set}
\format{
A list with components:
\describe{
  \item{CPIAUCSL}{inflation rate}
  \item{UNRATE}{unemployment rate}
  \item{GDPC1}{GDP growth rate}
}
}
\usage{
mf_usa
}
\description{
A dataset containing mixed-frequency data from FRED for three US macroeconomic variables.
}
\keyword{datasets}
