% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{mfbvar_sampler}
\alias{mfbvar_sampler}
\title{Posterior sampler for mfbvar}
\usage{
mfbvar_sampler(
  x,
  minn = FALSE,
  ssng = FALSE,
  ss = FALSE,
  dl = FALSE,
  csv = FALSE,
  fsv = FALSE,
  iw = FALSE,
  diffuse = FALSE,
  ...
)
}
\arguments{
\item{x}{Prior object of class \code{mfbvar_prior}}

\item{minn}{logical for \code{minn} prior}

\item{ssng}{logical for \code{ssng} prior}

\item{ss}{logical for \code{ss} prior}

\item{dl}{logical for \code{dl} prior}

\item{csv}{logical for \code{csv} variance structure}

\item{fsv}{logical for \code{fsv} variance structure}

\item{iw}{logical for \code{iw} variance prior}

\item{diffuse}{logical for \code{diffuse} variance prior}

\item{...}{additional arguments, currently only accepting lists \code{fixate}
and \code{init} with fixed values for certain parameters, or custom initial
values}
}
\value{
A list with the sampled parameters stored in arrays, and prior and
retrieved settings
}
\description{
Function that sets up the MCMC sampling routine for obtaining samples from
the posterior distribution
}
