% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar_schorf.R
\name{mfbvar_schorf}
\alias{mfbvar_schorf}
\title{MF-BVAR with a steady-state prior}
\usage{
mfbvar_schorf(Y, Lambda, prior_Pi_AR1, lambda1, lambda2, lambda3, n_lags,
  n_fcst, n_burnin, n_reps, verbose, ...)
}
\arguments{
\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}

\item{Lambda}{The Lambda matrix (size \code{n_vars* (n_vars*n_lags)}).}

\item{prior_Pi_AR1}{The prior means for the AR(1) coefficients.}

\item{lambda1}{The overall tightness.}

\item{lambda2}{The lag decay.}

\item{lambda3}{The tightness of the intercept prior variance.}

\item{n_lags}{The number of lags.}

\item{n_fcst}{The number of periods to forecast.}

\item{n_burnin}{The number of burn-in replications.}

\item{n_reps}{The number of replications.}

\item{verbose}{Logical, if progress should be printed to the console.}
}
\value{
An \code{mfbvar} object
}
\description{
The main function for running the MF-SS-BVAR.
}
\details{
\code{mfbvar} calls \code{\link{gibbs_sampler}} (implemented in C++)
}
\keyword{internal}
