% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{new_MFBVAR}
\alias{new_MFBVAR}
\title{Create a new object of the S3 class MFBVAR (the class name can be changed)}
\usage{
new_MFBVAR(...)
}
\description{
An object of the S3 class MFBVAR will be created inside the function.
}
\details{
The members (attributes) of the object give the settings for the mfbvar model.
The functions in the package take the object as the input.
The members of the object can be initialized by the function, and can also be changed by the following "set..." functions.
}
