% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols_pi}
\alias{ols_pi}
\alias{ols_s}
\title{OLS functions}
\usage{
ols_pi(X, Y)

ols_s(X, Y, Pi)
}
\arguments{
\item{X}{The regressor matrix.}

\item{Y}{The dependnet variable matrix.}

\item{Pi}{The estimated coefficients.
@keywords internal}
}
\value{
\item{pi_sample}{Estimated coefficients.}

\item{s_sample}{The sum of squared residuals matrix.}
}
\description{
Helper functions for multivariate regression and sum of squared error computations
}
\keyword{internal}
