% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_initialization}
\alias{parameter_initialization}
\title{Initialize parameters}
\usage{
parameter_initialization(
  Y,
  n_vars,
  n_lags,
  n_T_,
  init,
  n_fac = NULL,
  n_determ = NULL,
  n_sv = NULL,
  fsv,
  csv,
  ...
)
}
\arguments{
\item{Y}{data matrix}

\item{n_vars}{number of variables}

\item{n_lags}{number of lags}

\item{n_T_}{number of effective observations (lags subtracted)}

\item{init}{list with custom initial values}

\item{n_fac}{number of factors}

\item{n_determ}{number of determ}

\item{n_sv}{number of stochastic volatility series}

\item{fsv}{boolean for factor stochastic volatility}

\item{csv}{boolean for common stochastic volatility}
}
\description{
Function to initialize parameters (set starting values)
}
