% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{plot.mdd}
\alias{plot.mdd}
\title{Plotting method for class mdd}
\usage{
\method{plot}{mdd}(x, ...)
}
\arguments{
\item{x}{object of class mdd}

\item{...}{Currently not in use.}
}
\description{
Method for plotting mdd objects.
}
