% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{plot.mfbvar}
\alias{plot.mfbvar}
\title{Plotting method for class mfbvar}
\usage{
\method{plot}{mfbvar}(x, plot_start = NULL, ss_level = c(0.025, 0.975),
  pred_level = c(0.1, 0.9), ...)
}
\arguments{
\item{x}{object of class mfbvar}

\item{plot_start}{Time period (number) to start plotting from. Default is to to use \code{5*n_fcst} time periods if \code{n_fcst} exists, otherwise the entire sample.}

\item{ss_level}{A vector with the lower and upper quantiles for the posterior steady-state intervals.}

\item{pred_level}{A vector with the lower and upper quantiles for the forecast intervals.}

\item{...}{Currently not in use.}
}
\description{
Method for plotting mfbvar objects.
}
