% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot.mfbvar_prior}
\alias{plot.mfbvar_prior}
\title{Plot method for class \code{mfbvar_prior}}
\usage{
\method{plot}{mfbvar_prior}(x, nrow_facet = NULL, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_prior}}

\item{nrow_facet}{number of rows in facet}

\item{...}{Currently not in use.}
}
\description{
Method for plotting \code{mfbvar_prior} objects.
}
\details{
The function plots the data. If the prior moments for the steady-state parameters are available in \code{x}, these are included.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden[, 4:5], freq = c("m", "q"),
                       n_lags = 4, n_burnin = 20, n_reps = 20, n_fcst = 4)
plot(prior_obj)
}
