% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{print.mdd}
\alias{print.mdd}
\title{Printing method for class mdd}
\usage{
\method{print}{mdd}(x, ...)
}
\arguments{
\item{x}{object of class mdd}

\item{...}{Currently not in use.}
}
\description{
Method for printing mdd objects.
}
