% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{print.mfbvar}
\alias{print.mfbvar}
\title{Printing method for class mfbvar}
\usage{
\method{print}{mfbvar}(x, ...)
}
\arguments{
\item{x}{object of class mfbvar}

\item{...}{Currently not in use.}
}
\description{
Method for printing mfbvar objects.
}
