% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.mfbvar_minn}
\alias{print.mfbvar_minn}
\title{Printing method for class mfbvar_minn}
\usage{
\method{print}{mfbvar_minn}(x, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_minn}}

\item{...}{Currently not in use.}
}
\description{
Method for printing \code{mfbvar_minn} objects.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden[, 4:5], freq = c("m", "q"),
                       n_lags = 4, n_burnin = 20, n_reps = 20)
mod_minn <- estimate_mfbvar(prior_obj, prior_type = "minn")
print(mod_minn)
}
