% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{set_Member}
\alias{set_Member}
\title{Set attributes or members in an object of the class MFBVAR}
\usage{
set_Member(obj, ...)
}
\arguments{
\item{obj}{an object of the class MFBVAR}

\item{...}{the attributes and their values to be changed}
}
\description{
Set attributes or members in an object of the class MFBVAR
}
