\name{SimKF}
\alias{SimKF}
\title{Simulate}
\description{
Prints 'Hello, world!'.
}
\usage{
SimKF(mX, lH, mF, mB, mQ, iT, ip,
      iq, is, h0, P0=NULL, X0)
}
\arguments{
\item{mX}{}
\item{lH}{}
\item{mF}{}
\item{mB}{}
\item{mQ}{}
\item{iT}{}
\item{ip}{}
\item{iq}{}
\item{is}{}
\item{h0}{}
\item{P0}{}
\item{X0}{}
}
\details{}
\value{
\item{mZ}{}
\item{mH}{}
\item{me}{}
}

\examples{

}
