% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smoother}
\alias{smoother}
\alias{simulation_smoother}
\alias{generate_mhh}
\alias{loglike}
\alias{generate_mhh}
\alias{simulation_smoother}
\alias{loglike}
\title{Smooth and sample from the smoothed distribution}
\usage{
smoother(mZ, Lambda, mF, mQ, iT, ip, iq, h0, P0)

generate_mhh(mZ, Lambda, mF, mQ, iT, ip, iq, h0, P0)

simulation_smoother(mZ, Lambda, mF, mQ, iT, ip, iq, h0, P0)

loglike(mZ, Lambda, mF, mQ, iT, ip, iq, h0, P0)
}
\arguments{
\item{mZ}{\code{T * n_vars} matrix with the observations (\code{NA} represents missingness)}

\item{Lambda}{The Lambda matrix (size \code{n_vars* (n_vars*n_lags)}).}

\item{mF}{\code{(n_vars*n_lags) * (n_vars*n_lags)} matrix containing parameters (companion form)}

\item{mQ}{\code{(n_vars*n_lags) * (n_vars*n_lags)} matrix whose \code{n_vars*n_vars} top-left block is the Cholesky decomposition of the error covariance matrix}

\item{iT}{The sample size (sometimes called \code{n_T}).}

\item{ip}{The number of variables (\code{n_vars}).}

\item{iq}{The companion-form dimension (\code{n_vars*n_lags}).}

\item{h0}{The initial state (\code{(n_vars*n_lags)*1}).}

\item{P0}{The covariance matrix of the initial state (\code{(n_vars*n_lags)*(n_vars*n_lags)}).}
}
\value{
For \code{smoother}:
\item{}{The smoothed states.}

For \code{generate_mhh}:
\item{}{Generated (pseudo-)state vector.}

For \code{simulation_smoother}:
\item{}{The draw from the posterior distribution.}

For \code{loglike}:
\item{}{An \code{n_T}-long vector of the log-likelihoods. \code{exp(sum(loglike(...)))} is the likelihood.}
}
\description{
Functions for smoothing and sampling from the (smoothed) distribution \eqn{p(Z_{1:T}|Y_{1:T}, \Theta)}.
}
\details{
Implemented in C++.
}
\section{Functions}{
\itemize{
\item \code{smoother}: Compute smoothed states

\item \code{generate_mhh}: Generate pseudo-state vector

\item \code{simulation_smoother}: Simulation smoother

\item \code{loglike}: Compute log-likelihood
}}

