\name{smoothing}
\alias{smoothing}
\title{Compute the smoothed Kalman filter estimates}
\description{
Computes the smoothed estimates of \eqn{E(Z_{1:T}|Y_{1:T}, \Theta)}.
}
\usage{
smoothing(mZ, mX, lH, mF, mB, mQ, iT,
          ip, iq, is, h0, P0, X0)
}
\arguments{
\item{mZ}{}
\item{mX}{}
\item{lH}{}
\item{mF}{}
\item{mB}{}
\item{mQ}{}
\item{iT}{}
\item{ip}{}
\item{iq}{}
\item{is}{}
\item{h0}{}
\item{P0}{}
\item{X0}{}
}
\details{}
\value{
\item{mu}{}
}

\examples{

}
