% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{summary.mdd}
\alias{summary.mdd}
\title{Summary method for class mdd}
\usage{
\method{summary}{mdd}(object, ...)
}
\arguments{
\item{object}{object of class mdd}

\item{...}{Currently not in use.}
}
\description{
Method for summarizing mdd objects.
}
