% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{summary.mfbvar_minn}
\alias{summary.mfbvar_minn}
\title{Summary method for class \code{mfbvar_minn}}
\usage{
\method{summary}{mfbvar_minn}(object, ...)
}
\arguments{
\item{object}{object of class \code{mfbvar_minn}}

\item{...}{Currently not in use.}
}
\description{
Method for summarizing \code{mfbvar_minn} objects.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden[, 4:5], freq = c("m", "q"),
                       n_lags = 4, n_burnin = 20, n_reps = 20, n_fcst = 4)
mod_minn <- estimate_mfbvar(prior_obj, prior_type = "minn")
summary(mod_minn)

}
