% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{summary.mfbvar_prior}
\alias{summary.mfbvar_prior}
\title{Summary method for mfbvar_prior}
\usage{
\method{summary}{mfbvar_prior}(object, ...)
}
\arguments{
\item{object}{prior object (class \code{mfbvar_prior})}

\item{...}{additional arguments (currently unused)}
}
\description{
summary method for object of class mfbvar_prior, showing some basic
information regarding the contents of the prior.
}
\seealso{
\code{\link{set_prior}}, \code{\link{update_prior}}, \code{\link{estimate_mfbvar}}, \code{\link{print.mfbvar_prior}}
}
