% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{tidy.mfbvar}
\alias{tidy.mfbvar}
\title{Tidy method for class mfbvar}
\usage{
\method{tidy}{mfbvar}(x, result, ...)
}
\arguments{
\item{x}{object of class mfbvar}

\item{result}{string containing either \code{Pi}, \code{Sigma} or \code{psi} for which object to return}

\item{...}{Currently not in use.}
}
\description{
Method for obtaining tidy results from mfbvar objects.
}
\details{
The return can be large (especially for \code{result = "Pi"}).
}
