% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variable_initialization}
\alias{variable_initialization}
\title{Initialize variables}
\usage{
variable_initialization(Y, freq, freqs, n_lags, Lambda_, n_thin)
}
\arguments{
\item{Y}{data matrix}

\item{freq}{vector of frequencies for each column of \code{Y}}

\item{freqs}{vector with unique frequencies}

\item{n_lags}{number of lags}

\item{Lambda_}{aggregation matrix}

\item{n_thin}{thinning}
}
\description{
Function to initialize some variables used by samplers
}
