% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{Z_posterior}
\alias{Z_posterior}
\title{Draw from posterior of Z}
\usage{
Z_posterior(Y, d, Pi_r, Sigma_r, psi_r, Z_1, Lambda, n_vars, n_lags, n_T_,
  smooth_state)
}
\arguments{
\item{d}{The matrix of size \code{(n_T + n_lags) * n_determ} of deterministic terms.}

\item{n_vars}{The number of variables.}

\item{n_lags}{The number of lags.}
}
\value{
\code{Z_posterior} returns a list with:
\item{Z_r}{The draw of \code{Z}.}
\item{smoothed_Z_r}{(Only if \code{smooth_state == TRUE}) The smoothed state.}
}
\description{
Function for drawing from the posterior of Z using a simulation smoother, which can be used as a block in a Gibbs sampler.
}
