% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_companion}
\alias{build_companion}
\title{Build the companion matrix for the dynamic parameters}
\usage{
build_companion(Pi, n_vars, n_lags)
}
\arguments{
\item{Pi}{Matrix of size \code{n_vars * (n_vars*n_lags)} containing the dynamic coefficients.}

\item{n_vars}{The number of variables.}

\item{n_lags}{The number of lags.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
\item{Pi_comp}{The companion form matrix of size \code{(n_vars*n_lags) * (n_vars*n_lags)}}.
}
\description{
Builds the parameter matrix of dynamic coefficients for the companion form representation.
}
