% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_U}
\alias{build_U}
\title{Build the \code{U} matrix}
\usage{
build_U(Pi, n_determ)
}
\arguments{
\item{Pi}{Matrix of size \code{n_vars * (n_vars*n_lags)} containing the dynamic coefficients.}

\item{n_determ}{The number of deterministic terms.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
\item{U}{The \code{U} matrix, of size \code{((n_lags+1)n_vars*n_determ) * n_vars*n_determ}.}
}
\description{
Builds the parameter matrix of dynamic coefficients for the companion form representation.
}
