% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_Y_tilde}
\alias{build_Y_tilde}
\title{Build the lag-corrected data matrix}
\usage{
build_Y_tilde(Pi, z)
}
\arguments{
\item{Pi}{Matrix of size \code{n_vars * (n_vars*n_lags)} containing the dynamic coefficients.}

\item{z}{A matrix of size \code{(n_T + n_lags) * n_vars} of data.}
}
\value{
\item{Y_tilde}{A matrix of size \code{n_T * n_vars}.}
}
\description{
Builds the \eqn{\tilde{Y}=\Pi(L)Y} matrix.
}
\details{
Note that \code{z} does not contain missing values; at this point, the missing values have been replaced by values drawn using the simulation smoother.
}
