% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/densities.R
\name{rmatn}
\alias{rmatn}
\alias{rinvwish}
\alias{dnorminvwish}
\title{Normal inverse Wishart density function}
\usage{
rmatn(M, Q, P)

rinvwish(v, S)

dnorminvwish(X, Sigma, M, P, S, v)
}
\arguments{
\item{M}{The mean matrix of size \code{p * q}.}

\item{Q}{\code{q * q} covariance matrix.}

\item{P}{\code{p * p} covariance matrix.}

\item{v}{The degrees of freedom.}

\item{S}{\code{q * q} scale matrix.}

\item{X}{Matrix of size \code{p * q}.}

\item{Sigma}{The covariance matrix.}
}
\value{
For \code{dnorminvwish}: the evaluated density.\\n
For \code{rmatn} or \code{rinvwish}: the random numbers.
}
\description{
Density function for the (matrix) normal inverse Wishart distribution
}
