% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_mfbvar}
\alias{estimate_mfbvar}
\title{Mixed-frequency Bayesian VAR}
\usage{
estimate_mfbvar(mfbvar_prior = NULL, prior_type, ...)
}
\arguments{
\item{mfbvar_prior}{a \code{mfbvar_prior} object}

\item{prior_type}{either \code{"ss"} (steady-state prior) or \code{"minn"} (Minnesota prior)}

\item{...}{additional arguments to \code{update_prior} (if \code{mfbvar_prior} is \code{NULL}, the arguments are passed on to \code{set_prior})}
}
\value{
an object of class \code{mfbvar} and \code{mfbvar_ss}/\code{mfbvar_minn} containing posterior quantities as well as the prior object
}
\description{
The main function for estimating a mixed-frequency BVAR.
}
\details{
No (other than very primitive) checks are made on the additional arguments. Thus, for more informative error messages and a safer procedure, create the prior first using \code{set_prior}.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden, freq = c(rep("m", 4), "q"),
                       n_lags = 4, n_burnin = 20, n_reps = 20)
mod_minn <- estimate_mfbvar(prior_obj, prior_type = "minn")

}
\references{
Schorfheide, F., & Song, D. (2015) Real-Time Forecasting With a Mixed-Frequency VAR. \emph{Journal of Business & Economic Statistics}, 33(3), 366--380. \url{http://dx.doi.org/10.1080/07350015.2014.954707}\cr
Ankargren, S., Unosson, M., & Yang, Y. (2017) A Mixed-Frequency Bayesian Vector Autoregression with a Steady-State Prior. Unpublished manuscript.
}
\seealso{
\code{\link{set_prior}}, \code{\link{update_prior}}
}
