% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_mfbvar}
\alias{estimate_mfbvar}
\title{Mixed-frequency Bayesian VAR}
\usage{
estimate_mfbvar(mfbvar_prior = NULL, prior_type = c("ss", "minn"), ...)
}
\arguments{
\item{mfbvar_prior}{Test}

\item{prior_type}{Test 2}

\item{...}{test 3}
}
\description{
The main function for running the MF-SS-BVAR.
}
