% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_mfbvar}
\alias{estimate_mfbvar}
\title{Mixed-frequency Bayesian VAR}
\usage{
estimate_mfbvar(mfbvar_prior = NULL, prior, variance = "iw", ...)
}
\arguments{
\item{mfbvar_prior}{a \code{mfbvar_prior} object}

\item{prior}{either \code{"ss"} (steady-state prior) or \code{"minn"} (Minnesota prior)}

\item{variance}{form of the error variance-covariance matrix: \code{"iw"} for the inverse Wishart prior, \code{"diffuse"} for a diffuse prior, \code{"csv"} for common stochastic volatility or \code{"fsv"} for factor stochastic volatility}

\item{...}{additional arguments to \code{update_prior} (if \code{mfbvar_prior} is \code{NULL}, the arguments are passed on to \code{set_prior})}
}
\value{
An object of class \code{mfbvar}, \code{mfbvar_<prior>} and \code{mfbvar_<prior>_<variance>} containing posterior quantities as well as the prior object
}
\description{
The main function for estimating a mixed-frequency BVAR.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden, freq = c(rep("m", 4), "q"),
                       n_lags = 4, n_burnin = 20, n_reps = 20)
mod_minn <- estimate_mfbvar(prior_obj, prior = "minn")
}
\references{
Ankargren, S., Unosson, M., & Yang, Y. (2019) A Flexible Mixed-Frequency Bayesian Vector Autoregression with a Steady-State Prior. arXiv:????, \url{???}.\cr
Ankargren, S., & Jonéus, P. (2019) . arXiv:????, \url{???}.\cr
Kastner, G., & Huber, F. (2019) . arXiv:????, \url{???}.\cr
Schorfheide, F., & Song, D. (2015) Real-Time Forecasting With a Mixed-Frequency VAR. \emph{Journal of Business & Economic Statistics}, 33(3), 366--380. \url{http://dx.doi.org/10.1080/07350015.2014.954707}\cr
}
\seealso{
\code{\link{set_prior}}, \code{\link{update_prior}}, \code{\link{predict.mfbvar}}, \code{\link{plot.mfbvar_minn}},
\code{\link{plot.mfbvar_ss}}, \code{\link{varplot}}, \code{\link{summary.mfbvar}}
}
