% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fill_na}
\alias{fill_na}
\title{Fills NAs with the next non-NA value}
\usage{
fill_na(Y)
}
\arguments{
\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}
}
\value{
A matrix with no \code{NA}s.
}
\description{
The function fills elements with \code{NA} with the next non-\code{NA} value (so that quarterly averages observed at the end of the quarter are assumed as observations for the remaining months of the quarter).
}
