% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{mcmc_sampler}
\alias{mcmc_sampler}
\title{MCMC sampler}
\usage{
mcmc_sampler(x, ...)
}
\arguments{
\item{x}{argument to dispatch on (of class \code{prior_obj})}

\item{...}{additional named arguments passed on to the methods}
}
\description{
\code{mcmc_sampler} is a generic function for deciding which specific MCMC
algorithm to dispatch to. It is called internally.
}
