% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_minn}
\alias{mdd.mfbvar_minn}
\title{Marginal data density}
\usage{
\method{mdd}{mfbvar_minn}(x, quarterly_cols, type = "full", ...)
}
\description{
Estimate the marginal data density for the model with a Minnesota prior.
}
\details{
The generic \code{mdd} calls \code{mdd_minn} if \code{mfbvar_obj} is of class \code{mfbvar_minn}.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_ss}}
}
