% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_minn}
\alias{mdd.mfbvar_minn}
\title{Marginal data density method for class \code{mfbvar_minn}}
\usage{
\method{mdd}{mfbvar_minn}(x, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_minn}}

\item{...}{additional arguments (currently only \code{p_trunc} for the degree of truncation is available)}
}
\description{
Estimate the marginal data density for the model with a Minnesota prior.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_ss}}
}
