% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_minn}
\alias{mdd.mfbvar_minn}
\title{Marginal data density}
\usage{
\method{mdd}{mfbvar_minn}(x, quarterly_cols, type = c("full", "diff"), ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_minn}}

\item{quarterly_cols}{numeric vector with positions of quarterly variables}

\item{type}{\code{"full"} gives implementation based on the paper by Schorfheide and Song (2015), \code{"diff"} implementation on MATLAB code by Schorfheide and Song (2015)}

\item{...}{additional arguments (currently unused)}
}
\description{
Estimate the marginal data density for the model with a Minnesota prior.
}
\details{
The generic \code{mdd} calls \code{mdd_minn} if \code{mfbvar_obj} is of class \code{mfbvar_minn}.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_ss}}
}
