% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd.mfbvar_ss_iw}
\alias{mdd.mfbvar_ss_iw}
\title{Marginal data density method for class \code{mfbvar_ss}}
\usage{
\method{mdd}{mfbvar_ss_iw}(x, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar_ss}}

\item{...}{additional arguments (currently unused)}
}
\value{
The logarithm of the marginal data density.
}
\description{
Estimate the marginal data density for the model with a steady-state prior.
}
\details{
The marginal data density estimator follows Fuentes-Albero and Melosi (2013) and Ankargren, Unosson and Yang (2018).
}
\references{
Fuentes-Albero, C. and Melosi, L. (2013) Methods for Computing Marginal Data Densities from the Gibbs Output.
\emph{Journal of Econometrics}, 175(2), 132-141, \doi{10.1016/j.jeconom.2013.03.002}\cr
 Ankargren, S., Unosson, M., & Yang, Y. (2018) A Mixed-Frequency Bayesian Vector Autoregression with a Steady-State Prior. Working Paper, Department of Statistics, Uppsala University No. 2018:3.
}
\seealso{
\code{\link{mdd}}, \code{\link{mdd.mfbvar_minn_iw}}
}
