% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{estimate_mdd_minn}
\alias{estimate_mdd_minn}
\title{Estimate marginal data density in Minnesota MF-BVAR}
\usage{
estimate_mdd_minn(mfbvar_obj, p_trunc, ...)
}
\arguments{
\item{mfbvar_obj}{An object of class \code{mfbvar} containing the results.}

\item{p_trunc}{\code{1-p_trunc} is the degree of truncation (i.e. \code{p_trunc=1} is no truncation).}

\item{quarterly_cols}{numeric vector with positions of quarterly variables}
}
\value{
The log marginal data density estimate (bar a constant)
}
\description{
This function provides the possibility to estimate the log marginal density (up to a constant) using the Minnesota MF-BVAR.
}
\keyword{internal}
