% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mdd}
\alias{mdd}
\title{Marginal data density (method 2)}
\usage{
mdd(mfbvar_obj, p_trunc)
}
\arguments{
\item{p_trunc}{\code{1-p_trunc} is the degree of truncation (i.e. \code{p_trunc=1} is no truncation).}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}
}
\value{
A list with components being \code{n_reps}-long vectors. These can be used to estimate the MDD.
\item{pi_sigma_posterior}{Posterior of Pi and Sigma.}
\item{data_likelihood}{The likelihood.}
\item{pi_sigma_prior}{Prior of Pi and Sigma.}
\item{psi_prior}{Prior of psi.}
\item{psi_truncated}{The truncated psi pdf.}
}
\description{
Estimate the marginal data density using method 2.
}
