% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd.R
\name{mdd}
\alias{mdd}
\title{Marginal data density estimation}
\usage{
mdd(x, ...)
}
\arguments{
\item{x}{argument to dispatch on (of class \code{mfbvar_ss} or \code{mfbvar_minn})}

\item{...}{additional named arguments passed on to the methods}
}
\description{
\code{mdd} estimates the marginal data density.
}
\details{
This is a generic function. See the methods for more information.
}
\seealso{
\code{\link{mdd.mfbvar_ss}}, \code{\link{mdd.mfbvar_minn}}
}
