% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_schorf.R
\name{mdd_schorf}
\alias{mdd_schorf}
\title{Estimate MDD in Schorfheide-Song MFBVAR}
\usage{
mdd_schorf(mfbvar_obj, monthly_cols)
}
\arguments{
\item{monthly_cols}{Column indexes of monthly variables.}

\item{n_T}{The number of time points.}

\item{postsim}{The log marginal data density for \code{Z}.}

\item{Z}{The array of draws from the posterior of \code{Z}.}
}
\value{
The log marginal data density estimate (bar a constant)
}
\description{
This function provides the possibility to estimate the log marginal density (up to a constant) using the Schorfheide-Song MFBVAR.
}
