% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{mfbvar}
\alias{mfbvar}
\title{MF-BVAR with a steady-state prior}
\usage{
mfbvar(Y, d, d_fcst, Lambda, prior_Pi_AR1, lambda1, lambda2, prior_nu = NULL,
  prior_psi_mean, prior_psi_Omega, n_lags, n_fcst, n_burnin, n_reps, verbose)
}
\arguments{
\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}

\item{d}{The matrix of size \code{(n_T + n_lags) * n_determ} of deterministic terms.}

\item{d_fcst}{The deterministic terms for the forecasting period.}

\item{Lambda}{The Lambda matrix (size \code{n_vars* (n_vars*n_lags)}).}

\item{prior_Pi_AR1}{The prior means for the AR(1) coefficients.}

\item{lambda1}{The overall tightness.}

\item{lambda2}{The lag decay.}

\item{prior_nu}{The prior degrees of freedom.}

\item{prior_psi_mean}{Vector of length \code{n_determ*n_vars} with the prior means of the steady-state parameters.}

\item{prior_psi_Omega}{Matrix of size \code{(n_determ*n_vars) * (n_determ*n_vars)} with the prior covariance of the steady-state parameters.}

\item{n_lags}{The number of lags.}

\item{n_fcst}{The number of periods to forecast.}

\item{n_burnin}{The number of burn-in replications.}

\item{n_reps}{The number of replications.}

\item{verbose}{Logical, if progress should be printed to the console.}
}
\description{
The main function for running the MF-SS-BVAR.
}
\details{
\code{mfbvar} calls \code{\link{gibbs_sampler}} (implemented in C++)
}
