% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_psi}
\alias{posterior_psi}
\alias{posterior_psi_omega}
\title{Compute posterior moments of the steady-state parameters}
\usage{
posterior_psi(U, D, sigma, prior_psi_omega, psi_omega, Y_tilde, prior_psi)

posterior_psi_omega(U, D, sigma, prior_psi_omega)
}
\arguments{
\item{U}{\eqn{U} matrix, of size \code{(n_vars*n_determ*(n_lags+1)) * (n_vars*n_determ)}. This can be obtained using \code{\link{build_U}}.}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{sigma}{The covariance matrix.}

\item{prior_psi_omega}{Matrix of size \code{(n_determ*n_vars) * (n_determ*n_vars)} with the prior covariance of the steady-state parameters.}

\item{Y_tilde}{The lag-corrected data matrix (with no missing values) of size \code{n_T * n_vars}.}

\item{prior_psi}{Vector of length \code{n_determ*n_vars} with the prior means of the steady-state parameters.}
}
\value{
The return is:
\item{psi}{The posterior mean (from \code{\link{posterior_psi}})}

\item{psi_omega}{The posterior variance (from \code{\link{posterior_psi_omega}})}
}
\description{
Computes the mean and variance of the conditional posterior distribution of the steady-state parameters.
}
