% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{predict.mfbvar}
\alias{predict.mfbvar}
\title{Predict method for class mfbvar}
\usage{
\method{predict}{mfbvar}(object, pred_quantiles = c(0.1, 0.5, 0.9), ...)
}
\arguments{
\item{object}{object of class mfbvar}

\item{pred_quantiles}{The quantiles of the posterior predictive distribution to use.}

\item{...}{Currently not in use.}
}
\description{
The method for predicting mfbvar objects.
}
\details{
Note that this requires that forecasts were made in the original \code{mfbvar} call.
}
