% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.mfbvar}
\alias{print.mfbvar}
\title{Printing method for class mfbvar}
\usage{
\method{print}{mfbvar}(x, ...)
}
\arguments{
\item{x}{object of class \code{mfbvar}}

\item{...}{Currently not in use.}
}
\value{
No return value, called for side effects.
}
\description{
Method for printing \code{mfbvar} objects.
}
\examples{
prior_obj <- set_prior(Y = mf_usa, n_lags = 4, n_reps = 20)
mod_minn <- estimate_mfbvar(prior_obj, prior = "minn")
mod_minn
}
