% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pi_sigma.R
\name{prior_Pi_Sigma}
\alias{prior_Pi_Sigma}
\title{Create the priors for Pi and Sigma}
\usage{
prior_Pi_Sigma(lambda1, lambda2, prior_Pi_AR1, Y, n_lags, prior_nu)
}
\arguments{
\item{lambda1}{The overall tightness.}

\item{lambda2}{The lag decay.}

\item{prior_Pi_AR1}{The prior means for the AR(1) coefficients.}

\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}

\item{n_lags}{The number of lags.}

\item{prior_nu}{The prior degrees of freedom.}
}
\value{
\item{prior_Pi}{The prior mean matrix for Pi.}
\item{prior_Pi_Omega}{The prior covariance matrix for Pi.}
\item{prior_s}{The prior for Sigma.}
}
\description{
Creates the prior mean and covariance for Pi given the hyperparameters, and the prior parameters for Sigma.
}
