% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{psi_posterior}
\alias{psi_posterior}
\title{Draw from posterior of psi}
\usage{
psi_posterior(Pi_r, Sigma_r, Z_r1, prior_psi, prior_psi_omega, D, n_vars,
  n_lags, n_determ)
}
\arguments{
\item{Z_r1}{The previous draw of \code{Z} (i.e. \code{Z[,, r-1]}).}

\item{D}{The \code{D} matrix (from \code{\link{build_DD}}).}

\item{n_vars}{The number of variables.}

\item{n_lags}{The number of lags.}
}
\value{
\code{psi_posterior} returns:
\item{psi_r}{The draw of \code{psi}.}
}
\description{
Function for drawing from the posterior of psi, which can be used as a block in a Gibbs sampler.
}
