% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{set_prior}
\alias{set_prior}
\alias{update_prior}
\title{Set priors for an mfbvar model}
\usage{
set_prior(Y, d = NULL, d_fcst = NULL, freq, prior_Pi_AR1 = rep(0,
  ncol(Y)), lambda1 = 0.2, lambda2 = 1, lambda3 = 10000,
  prior_psi_mean = NULL, prior_psi_Omega = NULL, n_lags, n_fcst = 0,
  n_burnin, n_reps, verbose = FALSE, smooth_state = FALSE,
  check_roots = TRUE)

update_prior(prior_obj, ...)
}
\arguments{
\item{Y}{The data matrix of size \code{(n_T + n_lags) * n_vars} with \code{NA} representing missingness. Must correspond well to the specification of \code{M_Lambda}.}

\item{d}{The matrix of size \code{(n_T + n_lags) * n_determ} of deterministic terms.}

\item{d_fcst}{The deterministic terms for the forecasting period.}

\item{freq}{Character vector with elements 'm' (monthly) or 'q' (quarterly) for sampling frequency.}

\item{prior_Pi_AR1}{The prior means for the AR(1) coefficients.}

\item{lambda1}{The overall tightness.}

\item{lambda2}{The lag decay.}

\item{lambda3}{The tightness of the intercept prior variance.}

\item{prior_psi_mean}{Vector of length \code{n_determ*n_vars} with the prior means of the steady-state parameters.}

\item{prior_psi_Omega}{Matrix of size \code{(n_determ*n_vars) * (n_determ*n_vars)} with the prior covariance of the steady-state parameters.}

\item{n_lags}{The number of lags.}

\item{n_fcst}{The number of periods to forecast.}

\item{n_burnin}{The number of burn-in replications.}

\item{n_reps}{The number of replications.}

\item{verbose}{Logical, if progress should be printed to the console.}

\item{smooth_state}{Logical, if \code{TRUE} then the smoothed estimates of the latent states are also returned.}

\item{check_roots}{Logical, if roots of the companion matrix are to be checked to ensure stationarity.}

\item{prior_obj}{an object of class \code{mfbvar_prior}}

\item{...}{named arguments for prior attributes to update}
}
\description{
Create an object storing all information needed for estimation, including data as well as model and prior specifications for both a Minnesota or steady-state prior.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden, freq = c(rep("m", 4), "q"),
                       n_lags = 4, n_burnin = 100, n_reps = 100)
prior_obj <- update_prior(prior_obj, n_fcst = 4)

}
