% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbvar.R
\name{summary.mfbvar}
\alias{summary.mfbvar}
\title{Summary method for class mfbvar}
\usage{
\method{summary}{mfbvar}(object, ...)
}
\arguments{
\item{object}{object of class mfbvar}

\item{...}{Currently not in use.}
}
\description{
The method for summarizing mfbvar objects (currently the same as \code{print.mfbvar}).
}
