% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{summary.mfbvar}
\alias{summary.mfbvar}
\title{Summary method for class mfbvar}
\usage{
\method{summary}{mfbvar}(object, ...)
}
\arguments{
\item{object}{object of class \code{mfbvar}}

\item{...}{Currently not in use.}
}
\description{
Method for summarizing \code{mfbvar} objects.
}
\examples{
prior_obj <- set_prior(Y = mf_sweden[, 4:5], d = "intercept",
                       freq = c("m", "q"), n_lags = 4, n_burnin = 20, n_reps = 20)
mod_minn <- estimate_mfbvar(prior_obj, prior = "minn")
summary(mod_minn)
}
