% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{update_prior}
\alias{update_prior}
\title{Update priors for an mfbvar model}
\usage{
update_prior(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{mfbvar\_prior}}

\item{...}{named arguments for prior attributes to update}
}
\description{
Update priors for an mfbvar model
}
\seealso{
\code{\link{set_prior}}
}
